Introduce remote template cookbook
